requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Telegram Bot
python-telegram-bot==20.7

# LLM и AI
google-generativeai==0.3.2
openai==1.3.7

# Векторные базы данных
pinecone-client==2.2.4
chromadb==0.4.18

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Кеширование
redis==5.0.1
aioredis==2.0.1

# Парсинг и обработка данных
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
scrapy==2.11.0

# Утилиты
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Разработка
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Мониторинг
prometheus-client==0.19.0 